chore: delete message.service.ts, move data-access-heroes
