+++ demo/libs/tour-of-heroes/data-access-heroes/src/index.ts
@@ -1,4 +1,3 @@
 export * from './lib/data-access-heroes.module';
 
 export * from './lib/hero.service';
-export * from './lib/message.service';
