+++ demo/libs/tour-of-heroes/data-access-heroes/src/lib/hero.service.ts
