+++ demo/libs/tour-of-heroes/feature-hero-details/src/lib/feature-hero-details.module.ts
@@ -1,13 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HeroDetailComponent } from './hero-detail/hero-detail.component';
-import {
-  DataAccessHeroesModule,
-} from '@shared/data-access-heroes';
 import { RouterModule, Routes } from '@angular/router';
 import { FormsModule } from '@angular/forms';
+
 import { MatButtonModule } from '@angular/material/button';
 import { MatInputModule } from '@angular/material/input';
+
+import {DataAccessHeroesModule} from "@toh/data-access-heroes";
+import { HeroDetailComponent } from './hero-detail/hero-detail.component';
+
 import {
   MatSnackBarModule,
   MAT_SNACK_BAR_DEFAULT_OPTIONS,
