+++ demo/libs/tour-of-heroes/feature-hero-details/src/lib/hero-detail/hero-detail.component.ts
@@ -1,12 +1,13 @@
 import { Component } from '@angular/core';
 import {ActivatedRoute, ParamMap} from '@angular/router';
 import { Location } from '@angular/common';
+import { MatSnackBar } from '@angular/material/snack-bar';
+
+import { Observable } from 'rxjs';
+import { switchMap } from 'rxjs/operators';
 
 import { Hero } from '@shared/models';
-import { HeroService } from '@shared/data-access-heroes';
-import { Observable } from 'rxjs';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { switchMap } from 'rxjs/operators';
+import {HeroService} from "@toh/data-access-heroes";
 
 @Component({
   selector: 'app-hero-detail',
