+++ demo/libs/tour-of-heroes/feature-heroes/src/lib/feature-heroes.module.ts
@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HeroesComponent } from './heroes/heroes.component';
 import { RouterModule, Routes } from '@angular/router';
-import { DataAccessHeroesModule } from '@shared/data-access-heroes';
+
 import {MatButtonModule} from "@angular/material/button";
 import {MatIconModule} from "@angular/material/icon";
+
 import {FlipCardModule} from "@toh/shared-ui-kit";
+import {DataAccessHeroesModule} from "@toh/data-access-heroes";
+import { HeroesComponent } from './heroes/heroes.component';
 
 const routes: Routes = [
   { path: '', component: HeroesComponent },
