+++ demo/libs/tour-of-heroes/feature-heroes/src/lib/heroes/heroes.component.ts
@@ -1,8 +1,9 @@
 import {Component} from '@angular/core';
 
 import {Hero} from '@shared/models';
-import {HeroService} from '@shared/data-access-heroes';
+
 import {Observable} from 'rxjs';
+import {HeroService} from "@toh/data-access-heroes";
 
 @Component({
   selector: 'app-heroes',
